requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
affine==2.3.0
    # via rasterio
aiohttp==3.8.1
    # via pyvespa
aiosignal==1.2.0
    # via aiohttp
altair==4.1.0
    # via streamlit
argon2-cffi==21.1.0
    # via notebook
astor==0.8.1
    # via streamlit
async-timeout==4.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   environ-config
    #   fiona
    #   jsonschema
    #   rasterio
    #   streamlit
backcall==0.2.0
    # via ipython
backports.zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
base58==2.1.1
    # via streamlit
bleach==4.1.0
    # via nbconvert
blinker==1.4
    # via streamlit
boto3==1.20.8
    # via eodag
botocore==1.23.9
    # via
    #   boto3
    #   s3transfer
branca==0.4.2
    # via folium
cachetools==4.2.4
    # via streamlit
certifi==2021.10.8
    # via
    #   fiona
    #   pyproj
    #   rasterio
    #   requests
cffi==1.15.0
    # via
    #   argon2-cffi
    #   cryptography
charset-normalizer==2.0.7
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via
    #   click-plugins
    #   cligj
    #   eodag
    #   fiona
    #   flask
    #   rasterio
    #   streamlit
    #   usgs
click-plugins==1.1.1
    # via
    #   fiona
    #   rasterio
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
cryptography==36.0.0
    # via pyvespa
debugpy==1.5.1
    # via ipykernel
decorator==5.1.0
    # via
    #   ipdb
    #   ipython
    #   jsonpath-ng
    #   validators
defusedxml==0.7.1
    # via nbconvert
docker==5.0.3
    # via pyvespa
entrypoints==0.3
    # via
    #   altair
    #   jupyter-client
    #   nbconvert
environ-config==21.2.0
    # via -r requirements.in
eodag==2.3.4
    # via -r requirements.in
fiona==1.8.20
    # via geopandas
flasgger==0.9.5
    # via eodag
flask==2.0.2
    # via
    #   eodag
    #   flasgger
folium==0.12.1
    # via
    #   -r requirements.in
    #   streamlit-folium
frozenlist==1.2.0
    # via
    #   aiohttp
    #   aiosignal
geojson==2.5.0
    # via eodag
geopandas==0.10.2
    # via -r requirements.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.24
    # via streamlit
grpcio==1.31.0
    # via -r requirements.in
grpclib==0.4.2
    # via -r requirements.in
h2==4.1.0
    # via grpclib
hpack==4.0.0
    # via h2
hyperframe==6.0.1
    # via h2
idna==3.3
    # via
    #   requests
    #   yarl
importlib-metadata==4.8.2
    # via markdown
importlib-resources==5.4.0
    # via jsonschema
ipdb==0.13.9
    # via -r requirements.in
ipykernel==6.5.1
    # via
    #   ipywidgets
    #   notebook
    #   pydeck
ipython==7.29.0
    # via
    #   ipdb
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   nbformat
    #   notebook
ipywidgets==7.6.5
    # via pydeck
itsdangerous==2.0.1
    # via flask
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   altair
    #   branca
    #   flask
    #   folium
    #   nbconvert
    #   notebook
    #   pydeck
    #   pyvespa
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonpath-ng==1.5.3
    # via eodag
jsonschema==4.2.1
    # via
    #   altair
    #   flasgger
    #   nbformat
jupyter-client==7.0.6
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.9.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.2
    # via ipywidgets
lxml==4.6.4
    # via eodag
markdown==3.3.6
    # via eodag
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via
    #   flasgger
    #   nbconvert
more-itertools==8.11.0
    # via -r requirements.in
multidict==5.2.0
    # via
    #   aiohttp
    #   grpclib
    #   yarl
munch==2.5.0
    # via fiona
nbclient==0.5.8
    # via nbconvert
nbconvert==6.3.0
    # via notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.6
    # via widgetsnbextension
numpy==1.19.5
    # via
    #   -r requirements.in
    #   altair
    #   folium
    #   pandas
    #   pyarrow
    #   pydeck
    #   rasterio
    #   snuggs
    #   streamlit
owslib==0.25.0
    # via eodag
packaging==21.3
    # via
    #   bleach
    #   streamlit
pandas==1.3.4
    # via
    #   altair
    #   geopandas
    #   pyvespa
    #   streamlit
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.4.0
    # via streamlit
ply==3.11
    # via jsonpath-ng
prometheus-client==0.12.0
    # via notebook
prompt-toolkit==3.0.22
    # via ipython
protobuf==3.19.1
    # via streamlit
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pyarrow==6.0.1
    # via streamlit
pycparser==2.21
    # via cffi
pydeck==0.7.1
    # via streamlit
pygments==2.10.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pympler==0.9
    # via streamlit
pyparsing==3.0.6
    # via
    #   packaging
    #   snuggs
pyproj==3.3.0
    # via
    #   eodag
    #   geopandas
    #   owslib
pyrsistent==0.18.0
    # via jsonschema
pyshp==2.1.3
    # via eodag
pystac==1.2.0
    # via eodag
python-dateutil==2.8.2
    # via
    #   botocore
    #   eodag
    #   jupyter-client
    #   owslib
    #   pandas
    #   pystac
    #   streamlit
python-dotenv==0.19.2
    # via -r requirements.in
pytz==2021.3
    # via
    #   owslib
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyvespa==0.13.0
    # via -r requirements.in
pyyaml==6.0
    # via
    #   eodag
    #   flasgger
    #   owslib
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
rasterio==1.2.10
    # via -r requirements.in
requests==2.26.0
    # via
    #   docker
    #   eodag
    #   folium
    #   owslib
    #   pyvespa
    #   requests-futures
    #   streamlit
    #   usgs
requests-futures==1.0.0
    # via usgs
s3transfer==0.5.0
    # via boto3
send2trash==1.8.0
    # via notebook
shapely==1.8.0
    # via
    #   -r requirements.in
    #   eodag
    #   geopandas
six==1.16.0
    # via
    #   bleach
    #   fiona
    #   flasgger
    #   grpcio
    #   jsonpath-ng
    #   munch
    #   python-dateutil
    #   validators
smmap==5.0.0
    # via gitdb
snuggs==1.4.7
    # via rasterio
streamlit==1.2.0
    # via
    #   -r requirements.in
    #   streamlit-folium
streamlit-folium==0.4.0
    # via -r requirements.in
tenacity==8.0.1
    # via pyvespa
terminado==0.12.1
    # via notebook
testpath==0.5.0
    # via nbconvert
toml==0.10.2
    # via
    #   ipdb
    #   streamlit
toolz==0.11.2
    # via altair
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   streamlit
    #   terminado
tqdm==4.62.3
    # via
    #   -r requirements.in
    #   eodag
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   pydeck
typing-extensions==4.0.0
    # via
    #   async-timeout
    #   gitpython
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via streamlit
urllib3==1.26.7
    # via
    #   botocore
    #   requests
usgs==0.3.1
    # via eodag
validators==0.18.2
    # via streamlit
watchdog==2.1.6
    # via streamlit
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
websocket-client==1.2.1
    # via docker
werkzeug==2.0.2
    # via flask
whoosh==2.7.4
    # via eodag
widgetsnbextension==3.5.2
    # via ipywidgets
yarl==1.7.2
    # via aiohttp
zipp==3.6.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
